Renamed TabFragment1 to TopStoriesfragments         TabFragment2 to MostPopularfragment         TabFragment3 to Businessfragment         CellInformation to MostPopular
Supressed  in XML an useless toolbar in tab_fragment1,2 and 3
                  an item in menu_main, it was the hamburger button. The button is now in the upper left and works fine

In MainActivity, supressed extra useless code for the Tab and views pager.

In MostPopular, supressed the data that we didn't need.

Created a FragmentAdapter for displaying the fragment in our pager.

In InfoCellAdapter, adapted the adapter with the current data.

Created a class MostPopularResponse to get the data in MainActivity from the API.