In xml files : Modified some xml files to suit the models.
In mynews :
Finalised Businessfragment , TopStoriesfragment and MostPopularfragment.

In the models package :

Only change the name of some files and get rid of some useless stuff.

In package Utils:
Created the method for callbacks for the fragment TopStories and Business.
Added the call for TopStories and Business.

In adapter:
Finalized the adapter of our recycleview, everything is displayed as is should be