In adapter :
InfocellAdapter : Added a new method to display the webview

InfocellAdapterResearch : Create an adapter for the recycleview of the Research function.

       In models :

 JsonResponse : Added the short_url info, that's the url we need for the webview.

 JsonResponseResearch : created a new model for the json of the Research function

 ResearchResponse : created a response for the research function

      In utils :

CellInformationCalls : supressed some useless code

CellInformationCallsResearch : created a with retrofit for the research function

CellInformationService : added the call for the research function

      In java\com\miguel\mynews :

Filterfragment : cleaned some code with the shared preferences

Researchfragment : Added some method needed for the call with retrofit
For now, we got back an null response
Method researchFilterAndLaunchRetrofit create the list of the tags that the users clicked with the checkbox and launch the request with theses tags.

WebViewFragment : Created a fragment to display the webview when we clicked on recycleview cells

           In res\layout :

fragment_web_view: Created the layout of the fragment of our webviewFragment